update exceptions handling
