+++ src/main/java/com/sourcery/gymapp/backend/workout/exception/ErrorResponse.java
@@ -1,16 +1,7 @@
 package com.sourcery.gymapp.backend.workout.exception;
 
-import java.util.List;
-
 public record ErrorResponse(
         String message,
-        ErrorCode code,
-        List<FieldResponse> fields
-) {
-}
-
-record FieldResponse(
-        String field,
-        String error
+        ErrorCode code
 ) {
 }
