+++ src/main/java/com/sourcery/gymapp/backend/workout/exception/FieldErrorResponse.java
@@ -1,0 +1,16 @@
+package com.sourcery.gymapp.backend.workout.exception;
+
+import java.util.List;
+
+public record FieldErrorResponse(
+        String message,
+        ErrorCode code,
+        List<FieldResponse> fields
+) {
+}
+
+record FieldResponse(
+        String field,
+        String error
+) {
+}
